refactor merge register & login domain
