+++ app/src/main/java/com/traveloka/hotel/common/data/MainRepository.kt
@@ -6,6 +6,8 @@ import com.traveloka.hotel.common.data.local.IUserPreference
 class MainRepository(private val userPreference: IUserPreference) {
     fun setLocation(location: LatLng) = userPreference.setLocation(location)
     fun getLocation() = userPreference.getLocation()
+    fun getToken() = userPreference.fetchAuthToken()
+    fun setToken(token: String) = userPreference.saveAuthToken(token)
 
     companion object {
         @Volatile
