+++ app/src/main/java/com/traveloka/hotel/common/data/local/UserPreference.kt
@@ -4,10 +4,14 @@ import android.content.Context
 import com.google.android.gms.maps.model.LatLng
 
 interface IUserPreference {
-    fun setLocation(value: LatLng)
+    fun setLocation(value: LatLng?)
     fun getLocation(): LatLng
     fun saveAuthToken(token: String)
     fun fetchAuthToken(): String?
+    fun saveEmail(token: String)
+    fun getEmail(): String?
+    fun savePassword(token: String)
+    fun getPassword(): String?
 }
 
 
@@ -17,14 +21,16 @@ class UserPreference(context: Context) : IUserPreference {
         private const val LATITUDE = "latitude"
         private const val LONGITUDE = "longitude"
         const val USER_TOKEN = "user_token"
+        const val USER_EMAIL = "user_email"
+        const val USER_PASSWORD = "user_password"
     }
 
     private val preferences = context.getSharedPreferences(PREFS_NAME, Context.MODE_PRIVATE)
 
-    override fun setLocation(value: LatLng) {
+    override fun setLocation(value: LatLng?) {
         val editor = preferences.edit()
-        editor.putString(LATITUDE, value.latitude.toString())
-        editor.putString(LONGITUDE, value.longitude.toString())
+        editor.putString(LATITUDE, value?.latitude?.toString() ?: "")
+        editor.putString(LONGITUDE, value?.longitude?.toString() ?: "")
         editor.apply()
     }
 
@@ -44,6 +50,26 @@ class UserPreference(context: Context) : IUserPreference {
 
     override fun fetchAuthToken(): String? {
         return preferences.getString(USER_TOKEN, null)
+    }
+
+    override fun saveEmail(token: String) {
+        val editor = preferences.edit()
+        editor.putString(USER_EMAIL, token)
+        editor.apply()
+    }
+
+    override fun getEmail(): String? {
+        return preferences.getString(USER_EMAIL, null)
+    }
+
+    override fun savePassword(token: String) {
+        val editor = preferences.edit()
+        editor.putString(USER_PASSWORD, token)
+        editor.apply()
+    }
+
+    override fun getPassword(): String? {
+        return preferences.getString(USER_PASSWORD, null)
     }
