+++ app/src/main/java/com/traveloka/hotel/common/domain/MainViewModel.kt
@@ -13,4 +13,6 @@ class MainViewModel(private val mainRepository: MainRepository) : ViewModel() {
     }
 
     fun getLocation() = mainRepository.getLocation()
+    fun getToken() = mainRepository.getToken()
+    fun setToken(token: String) = mainRepository.setToken(token)
 }
