+++ app/src/main/java/com/traveloka/hotel/common/presentation/MainActivity.kt
@@ -3,7 +3,7 @@ package com.traveloka.hotel.common.presentation
 import android.os.Bundle
 import androidx.activity.ComponentActivity
 import androidx.activity.compose.setContent
-import com.traveloka.hotel.ui.navigation.HotelNavigation
+import com.traveloka.hotel.common.presentation.navigation.HotelNavigation
 import com.traveloka.hotel.ui.theme.HotelmobileTheme
 
 class MainActivity : ComponentActivity() {
