+++ app/src/main/java/com/traveloka/hotel/common/presentation/navigation/HotelNavigation.kt
@@ -1,4 +1,4 @@
-package com.traveloka.hotel.ui.navigation
+package com.traveloka.hotel.common.presentation.navigation
 
 import androidx.compose.runtime.Composable
 import androidx.navigation.NavType
@@ -6,10 +6,9 @@ import androidx.navigation.compose.NavHost
 import androidx.navigation.compose.composable
 import androidx.navigation.compose.rememberNavController
 import androidx.navigation.navArgument
-import com.traveloka.hotel.common.presentation.navigation.HotelScreens
 import com.traveloka.hotel.common.presentation.screens.introduction.IntroScreen
-import com.traveloka.hotel.common.presentation.screens.login.LoginScreen
-import com.traveloka.hotel.common.presentation.screens.register.RegisterScreen
+import com.traveloka.hotel.featureAuth.presentation.login.LoginScreen
+import com.traveloka.hotel.featureAuth.presentation.register.RegisterScreen
 import com.traveloka.hotel.featureHotel.presentation.detailHotel.DetailHotelScreen
 import com.traveloka.hotel.featureHotel.presentation.listHotel.ListHotelScreen
 import com.traveloka.hotel.featureReview.presentation.listReview.ListReviewScreen
