--- app/src/main/java/com/traveloka/hotel/common/util/WrapperFunc.kt
@@ -1,45 +1,0 @@
-package com.traveloka.hotel.common.util
-
-import androidx.compose.runtime.Composable
-import androidx.compose.runtime.LaunchedEffect
-import androidx.compose.ui.platform.LocalContext
-import androidx.navigation.NavController
-import com.traveloka.hotel.common.data.local.UserPreference
-import com.traveloka.hotel.common.presentation.navigation.HotelScreens
-
-object WrapperFunc {
-
-    @Composable
-    fun WithAuth(navController: NavController, content: @Composable () -> Unit) {
-        val context = LocalContext.current
-        val userPreference = UserPreference(context)
-
-        LaunchedEffect(userPreference.fetchAuthToken()) {
-            if (userPreference.fetchAuthToken().isNullOrEmpty()) {
-                navController.navigate(HotelScreens.LoginScreen.name)
-            }
-        }
-
-        if (!userPreference.fetchAuthToken().isNullOrEmpty()) {
-            content()
-        }
-
-    }
-
-    @Composable
-    fun WithNonAuth(navController: NavController, content: @Composable () -> Unit) {
-        val context = LocalContext.current
-        val userPreference = UserPreference(context)
-
-        LaunchedEffect(userPreference.fetchAuthToken()) {
-            if (!userPreference.fetchAuthToken().isNullOrEmpty()) {
-                navController.navigate(HotelScreens.ListHotelScreen.name)
-            }
-        }
-
-        if (userPreference.fetchAuthToken().isNullOrEmpty()) {
-            content()
-        }
-
-    }
-}
