+++ app/src/main/java/com/traveloka/hotel/featureAuth/presentation/register/RegisterScreen.kt
@@ -1,4 +1,4 @@
-package com.traveloka.hotel.common.presentation.screens.register
+package com.traveloka.hotel.featureAuth.presentation.register
 
 import androidx.compose.foundation.background
 import androidx.compose.foundation.layout.*
@@ -11,10 +11,10 @@ import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import androidx.navigation.NavController
 import androidx.navigation.compose.rememberNavController
-import com.traveloka.hotel.common.util.WithLocation
-import com.traveloka.hotel.common.util.WrapperFunc.WithNonAuth
-import com.traveloka.hotel.featureRegister.presentation.components.RegisterBody
-import com.traveloka.hotel.featureRegister.presentation.components.RegisterHeader
+import com.traveloka.hotel.component.WithLocation
+import com.traveloka.hotel.component.WrapperFunc.WithNonAuth
+import com.traveloka.hotel.featureAuth.presentation.register.components.RegisterBody
+import com.traveloka.hotel.featureAuth.presentation.register.components.RegisterHeader
 import com.traveloka.hotel.ui.theme.BlueBg
 import com.traveloka.hotel.ui.theme.HotelmobileTheme
