+++ app/src/main/java/com/traveloka/hotel/featureAuth/presentation/register/components/RegisterForm.kt
@@ -1,4 +1,4 @@
-package com.traveloka.hotel.featureLogin.presentation.components
+package com.traveloka.hotel.featureAuth.presentation.register.components
 
 import android.widget.Toast
 import androidx.compose.foundation.layout.Arrangement
@@ -13,37 +13,37 @@ import androidx.compose.ui.Modifier
 import androidx.compose.ui.platform.LocalContext
 import androidx.compose.ui.platform.LocalSoftwareKeyboardController
 import androidx.compose.ui.res.stringResource
-import androidx.compose.ui.text.input.ImeAction
-import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import androidx.lifecycle.viewmodel.compose.viewModel
 import androidx.navigation.NavController
-import androidx.navigation.compose.rememberNavController
 import com.traveloka.hotel.R
 import com.traveloka.hotel.common.data.ResultApi
-import com.traveloka.hotel.common.data.local.UserPreference
 import com.traveloka.hotel.common.presentation.navigation.HotelScreens
-import com.traveloka.hotel.common.presentation.screens.login.LoginScreen
 import com.traveloka.hotel.component.button.MButton
+import com.traveloka.hotel.component.form.DateField
 import com.traveloka.hotel.component.form.EmailField
 import com.traveloka.hotel.component.form.PasswordField
-import com.traveloka.hotel.featureLogin.data.model.LoginRequest
-import com.traveloka.hotel.featureLogin.domain.LoginViewModel
-import com.traveloka.hotel.featureLogin.domain.LoginViewModelFactory
-import com.traveloka.hotel.ui.theme.HotelmobileTheme
+import com.traveloka.hotel.component.form.RadioField
+import com.traveloka.hotel.featureAuth.data.model.register.RegisterRequest
+import com.traveloka.hotel.featureAuth.domain.AuthViewModel
+import com.traveloka.hotel.featureAuth.domain.AuthViewModelFactory
 
 @OptIn(ExperimentalComposeUiApi::class)
 @Composable
-fun LoginForm(
-    navController: NavController,
-    viewModel: LoginViewModel = viewModel(factory = LoginViewModelFactory.getInstance(LocalContext.current))
+fun RegisterForm(
+    viewModel: AuthViewModel = viewModel(
+        factory = AuthViewModelFactory.getInstance(LocalContext.current)
+    ),
+    navController: NavController
 ) {
+    val options = listOf(stringResource(R.string.male), stringResource(R.string.female))
     val mContext = LocalContext.current
 
-    val loginState = viewModel.loginState
+    val registerState = viewModel.registerState
     val emailState = viewModel.email
     val passwordState = viewModel.password
-
+    val genderState = viewModel.gender
+    val birthDateState = viewModel.birthDate
     val keyboardController = LocalSoftwareKeyboardController.current
 
     var isSubmitEnabled by remember {
@@ -52,37 +52,36 @@ fun LoginForm(
     var isLoading by remember {
         mutableStateOf(false)
     }
-    val handleLogin = {
-        viewModel.login(
-            LoginRequest(
+    val handleRegister = {
+        viewModel.register(
+            RegisterRequest(
                 email = emailState.value,
-                password = passwordState.value
+                password = passwordState.value,
+                gender = genderState.value,
+                birthDate = birthDateState.value
             )
         )
     }
 
-    LaunchedEffect(emailState.value, passwordState.value) {
-        isSubmitEnabled = emailState.value.isNotBlank() && passwordState.value.isNotBlank()
-    }
-
-    LaunchedEffect(loginState.value) {
-        when (val loginStateVal = loginState.value) {
+    LaunchedEffect(registerState.value) {
+        when (val registerStateVal = registerState.value) {
             is ResultApi.Success -> {
                 isSubmitEnabled = true
                 isLoading = false
-                val token = loginStateVal.data
-                val userPreference = UserPreference(mContext)
-                if (!token.isNullOrEmpty()) {
-                    userPreference.saveAuthToken(token)
-                }
-                navController.navigate(HotelScreens.ListHotelScreen.name)
+                viewModel.clearForm()
+                Toast.makeText(
+                    mContext,
+                    registerStateVal.data,
+                    Toast.LENGTH_LONG
+                ).show()
+                navController.navigate(HotelScreens.LoginScreen.name)
             }
             is ResultApi.Failure -> {
                 isSubmitEnabled = true
                 isLoading = false
                 Toast.makeText(
                     mContext,
-                    loginStateVal.error,
+                    registerStateVal.error,
                     Toast.LENGTH_LONG
                 ).show()
             }
@@ -92,6 +91,12 @@ fun LoginForm(
             }
             else -> {}
         }
+
+    }
+
+    LaunchedEffect(emailState.value, passwordState.value, genderState.value, birthDateState.value) {
+        isSubmitEnabled =
+            emailState.value.isNotBlank() && passwordState.value.isNotBlank() && genderState.value.isNotBlank() && birthDateState.value.isNotBlank()
     }
 
     Column(
@@ -112,27 +117,28 @@ fun LoginForm(
             onValueChange = { viewModel.setPassword(it) },
             placeholder = stringResource(R.string.password_hint),
             title = stringResource(R.string.password),
-            imeAction = ImeAction.Done,
             keyboardActions = KeyboardActions(onDone = {
                 keyboardController?.hide()
             })
         )
-
+        RadioField(
+            title = stringResource(R.string.gender),
+            options = options,
+            selectedItem = genderState.value,
+            onValueChange = { newGender -> viewModel.setGender(newGender) }
+        )
+        DateField(
+            onValueChange = { viewModel.setBirthDate(it) },
+            placeholder = "yyyy-mm-dd",
+            title = stringResource(R.string.birth_date)
+        )
 
         MButton(
-            enabled = isSubmitEnabled,
             modifier = Modifier.widthIn(min = 200.dp),
+            enabled = isSubmitEnabled,
             isLoading = isLoading,
-            text = stringResource(R.string.log_in),
-            onClick = handleLogin
+            text = stringResource(R.string.register),
+            onClick = handleRegister
         )
-    }
-}
-
-@Preview(showSystemUi = true)
-@Composable
-fun LoginFormPreview() {
-    HotelmobileTheme {
-        LoginScreen(rememberNavController())
     }
 }
