+++ app/src/main/java/com/traveloka/hotel/featureAuth/presentation/register/components/RegisterHeader.kt
@@ -1,4 +1,4 @@
-package com.traveloka.hotel.featureRegister.presentation.components
+package com.traveloka.hotel.featureAuth.presentation.register.components
 
 import androidx.compose.foundation.layout.Row
 import androidx.compose.foundation.layout.fillMaxWidth
