+++ app/src/main/java/com/traveloka/hotel/featureHotel/presentation/detailHotel/DetailHotelScreen.kt
@@ -23,7 +23,7 @@ import coil.compose.AsyncImage
 import coil.request.ImageRequest
 import com.traveloka.hotel.R
 import com.traveloka.hotel.common.presentation.navigation.HotelScreens
-import com.traveloka.hotel.common.util.WrapperFunc.WithAuth
+import com.traveloka.hotel.component.WrapperFunc.WithAuth
 import com.traveloka.hotel.featureHotel.presentation.detailHotel.components.Content
 import com.traveloka.hotel.featureHotel.presentation.detailHotel.components.Reviews
 import com.traveloka.hotel.featureHotel.util.getHotels
