--- app/src/main/java/com/traveloka/hotel/featureRegister/domain/RegisterViewModelFactory.kt
@@ -1,29 +1,0 @@
-package com.traveloka.hotel.featureRegister.domain
-
-import android.content.Context
-import androidx.lifecycle.ViewModel
-import androidx.lifecycle.ViewModelProvider
-import com.traveloka.hotel.di.Injection
-import com.traveloka.hotel.featureRegister.data.RegisterRepository
-
-class RegisterViewModelFactory private constructor(private val registerRepository: RegisterRepository) :
-    ViewModelProvider.NewInstanceFactory() {
-
-    @Suppress("UNCHECKED_CAST")
-    override fun <T : ViewModel> create(modelClass: Class<T>): T {
-        if (modelClass.isAssignableFrom(RegisterViewModel::class.java)) {
-            return RegisterViewModel(registerRepository) as T
-        }
-        throw IllegalArgumentException("Unknown ViewModel class")
-    }
-
-    companion object {
-        @Volatile
-        private var instance: RegisterViewModelFactory? = null
-
-        fun getInstance(context: Context): RegisterViewModelFactory =
-            instance ?: synchronized(this) {
-                instance ?: RegisterViewModelFactory(Injection.provideRegisterRepository(context))
-            }.also { instance = it }
-    }
-}
