+++ app/src/main/res/values/strings.xml
@@ -34,4 +34,5 @@
     <string name="welcome_back">Welcome back</string>
     <string name="lets_explore_the_world">Let\'s explore the world</string>
     <string name="dont_have_account">Don’t have an account?</string>
+    <string name="token_not_provided">Token wasn\'t provided</string>
 </resources>
