+++ build.gradle
@@ -7,11 +7,6 @@ buildscript {
         google()
         mavenCentral()
     }
-    dependencies {
-        classpath "com.android.tools.build:gradle:7.0.4"
-        classpath "org.jetbrains.kotlin:kotlin-gradle-plugin:1.5.21"
-        classpath "com.google.dagger:hilt-android-gradle-plugin:2.38.1"
-    }
 }
 
 // Top-level build file where you can add configuration options common to all sub-projects/modules.
